Calculating the distance between two coordinates
Calculating the distance between two coordinates on the ride through an api call to OSM

+++ taxi-app/backend/services/ride-service/src/main/java/com/taxi/ride/dto/RideDto.java
@@ -42,5 +42,5 @@ public class RideDto {
     private double endLatitude;
     private double endLongitude;
 
-    private BigDecimal distance;
+    private double distance;
 }
