+++ taxi-app/backend/services/ride-service/src/main/java/com/taxi/ride/entities/Ride.java
@@ -39,5 +39,5 @@ public class Ride {
     private double endLatitude;
     private double endLongitude;
 
-    private BigDecimal distance;
+    private double distance;
 }
